Separated Session, Player, and GameRoom Components
