+++ game_room.py
@@ -1,68 +1,58 @@
-from game_logic import NumberGame
+from game_errors import *
+from game_logic import *
 
+def get_turn_obj(data, game):
+    if data["type"] == "draw":
+        turn_obj = Turn_Draw(game)
+    elif data["type"] == "attack":
+        turn_obj = Turn_Attack(game, 
+                              data["cards"])
+    else:
+        raise SelectionError("To draw, only select the deck") # TODO # implement this into javascript
+    return turn_obj
 
-GAME_ROOMS = dict()
-
-class Player:
-    def __init__(self, sid, name):
-        self.__sid = sid
-        self.name = name
-        self.game_room = None
-
-    def create_game(self, opponent_name, game_name):
-        game = NumberGame(self.name, opponent_name)
-        room = GameRoom(game)
-        GAME_ROOMS[game_name] = room
-
-    def join_game(self, game_name):
-        game_room = GAME_ROOMS[game_name]
-        self.game_room = game_room
-        self.game_room.update_state()
-
-    def update_state(self, state):
-        emit('set_state', state, to=self.sid)
-
-    def submit_turn(self, turn):
-        self.game_room.do_turn(self, turn)
-
-    def validate_turn(self, data):
-        self.game_room.check_turn(self, turn)
-
-    def leave_room(self):
-        if self.game_room:
-            self.game_room.remove_player(self)
-            self.game_room = None
 
 class GameRoom:
-    def __init__(self, game):
-        self.players = dict()
-        self.game = game
+    def __init__(self, game, name):
+        self.__game = game
+        self.__room_id = name
+        self.__players = list()
         
+    def validate_turn(self, player, data):
+        if self.__game.get_active_player() != player.name:
+            raise SelectionError("Not Your Turn")
+        turn_obj = get_turn_obj(data, self.__game)
+        status = turn_obj.validate()
+        return status
+
+    def do_turn(self, player, data):
+        if self.__game.get_active_player() != player.name:
+            raise SelectionError("Not Your Turn")
+        turn_obj = get_turn_obj(data, self.__game)
+        status = turn_obj.execute()
+        self.update_players()
+
     def add_player(self, player):
-        self.players[player.name] = player
+        self.__players.append(player)
 
     def remove_player(self, player):
-        del self.players[player.name]
+        self.__players.remove(player)
 
-    def update_state(self):
-        state = self.game.get_state()
-        active_player = self.game.get_active_player()
-        score = self.game.get_score()
+    def update_players(self):
+        game = self.__game
+        state = game.get_state()
+        for player in self.__players:
+            player.set_state(state)
 
-        state = {"cards": state, 
-                 "active_player": active_player, 
-                 "score": score}
-        for player in self.players.items():
-            player.update_state(state)
+    def get_players(self):
+        game = self.__game
+        return [game.get_active_player(), game.get_inactive_player()]
 
-    def submit_turn(self, player, data):
-        if player.name != self.game.get_active_player():
-            raise SelectionError("Not Your Turn")
-        self.turn.validate()
-        self.update_state()
+    def get_state(self):
+        game = self.__game
+        state = game.get_state()
+        return state
 
-    def validate_turn(self, player, data):
-        if player.name != self.game.get_active_player():
-            raise SelectionError("Not Your Turn")
-        self.turn.execute()
+
+
